Add test for calculators

+++ src/main/java/usecase_pokemon/ExperiencePointCalculator.java
@@ -4,6 +4,6 @@ import entity.Pokemon;
 
 public class ExperiencePointCalculator {
     public int calculate(Pokemon pokemon) {
-        return pokemon.getExperiencePoint() / 5;
+        return (int) (pokemon.getExperiencePoint() / 5.0);
     }
 }
